reorganize tools
